Highlight GPT model button when active
Added a highlight class to the button responsible for switching between GPT 3 and GPT 4 in the chat window. The "highlight" CSS class changes the color of the active GPT model icon when GPT 4 is being used, to make it easier for the user to identify the currently applied model at a glance.